requirements.txt
flask==3.0.3
flask-cors==4.0.0
mysql-connector-python==8.3.0
numpy==1.26.4
werkzeug==3.0.3
opencv-python==4.9.0.80
face_recognition==1.3.0
